image-fixer/requirements.txt
Keras==2.3.1
tensorflow==2.1.0
numpy==1.18.1
matplotlib==3.1.3
Pillow==2.2.1